+++ CHANGES.md
@@ -1,4 +1,6 @@
 #### [unreleased]
+* add most of move_dir() and is_virtualbox()
+* use GU_Trait::copy_dir() which is recursive copy > unlink, core copy_dir() doesn't work with how Git Updater renames downloads.
 
 #### 10.9.0 / 2022-04-04
 * revert usage of `move_dir()` and `is_virtualbox()` -- for now