+++ CHANGES.md
@@ -1,4 +1,6 @@
 #### [unreleased]
+
+#### 10.10.0 / 2022-04-23
 * add most of `move_dir()` and `is_virtualbox()`
 * update above to work with Rollback
 * put `move_dir()` and `is_virtualbox()` in `class Shim` for simpler usage from core functions